Distance Matrix API working properly. Added static class for converting meters to miles

+++ Models/Owner.cs
@@ -35,7 +35,7 @@ namespace PawMates.Models
         [Display(Name = "Filter Temperment")]
         public string? FilterTemperment { get; set; }
         [Display(Name = "Filter Distance")]
-        public int? FilterDistance { get; set; }
+        public double? FilterDistance { get; set; }
         public double? OwnerLatitude { get; set; }
         public double? OwnerLongitude { get; set; }
